requirements.txt
chroma
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchainhub
langdetect
langsmith
fastapi
fastapi-cli
requests
SQLAlchemy
tiktoken
uvicorn
uvloop
python-decouple
pypdf

